fix: Attachment button functionality and platform-specific shortcuts
- Move file input handling from ImageAttachment to MessageForm for cleaner architecture
- Fix attachment button rendering and click functionality
- Add platform detection for keyboard shortcuts (Cmd+V on macOS, Ctrl+V elsewhere)
- Simplify component prop passing by moving file logic closer to the UI
- Ensure attachment button appears inline with textarea and send button
- Remove complex callback patterns that caused immediate file dialog opening
